move serve command to archiver.serve
